Move interface compute tags into separate file
